pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "voice-ui"
version = "0.1.0"
dependencies = [
  "colorama",
  "pveagle",
  "pvporcupine",
  "pvrecorder",
  "pyaudio",
  "python-dotenv",
  "requests",
  "six",
  "sox",
  "pydub",
]
requires-python = ">=3.8"
authors = [
  {name = "Anderson Silva", email = "andyluiz@yahoo.com"},
]
maintainers = [
  {name = "Anderson Silva", email = "andyluiz@yahoo.com"}
]
description = "Voice UI supporting transcription, real-time exchange and speech interruption. This is meant to enable voice support to applications using large language models."
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["voice", "UI", "LLM"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: MIT License",
  "Intended Audience :: Developers",
  "Programming Language :: Python",
]

[project.optional-dependencies]
openai = [
  "openai",
]
google = [
  "google-cloud-speech",
  "google-cloud-texttospeech",
]
local-whisper = [
  "whisper_timestamped",
]
silero = [
  "silero-vad",
  "torch>=1.12.0",
  "torchaudio>=0.12.0",
]
pvcobra = [
  "pvcobra",
]
funasr = [
  "funasr",
  "torch>=1.13",
  "torchaudio",
]
webrtc = [
  "aiortc>=1.5.0",
  "av>=10.0.0",
  "websockets>=11.0",
]
test = [
  "coverage",
]
dev = [
  "black",
  "ruff",
  "mypy",
  "isort",
]

[tool.black]
preview = false

[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
per-file-ignores = { "tests/*" = ["E402"] }
